+++ src/lib/backends/core-lightning/utils.ts
@@ -1,6 +1,6 @@
 import Big from 'big.js'
 import type { Payment } from '$lib/types'
-import { decodeBolt11, formatMsat, logger } from '$lib/utils'
+import { decodeBolt11, formatDecodedOffer, formatMsat, logger } from '$lib/utils'
 import lightning from '$lib/lightning'
 
 import type {
@@ -11,6 +11,7 @@ import type {
   IncomeEvent,
   DecodedBolt12Invoice
 } from './types'
+import { decodedOffers$ } from '$lib/streams'
 
 export function invoiceStatusToPaymentStatus(status: InvoiceStatus): Payment['status'] {
   switch (status) {
@@ -58,6 +59,7 @@ export async function invoiceToPayment(invoice: Invoice): Promise<Payment> {
 
   if (bolt12) {
     const decoded = await lnApi.decode(bolt12)
+
     const { invoice_created_at, offer_issuer, offer_id, invreq_payer_note } =
       decoded as DecodedBolt12Invoice
 
@@ -69,6 +71,9 @@ export async function invoiceToPayment(invoice: Invoice): Promise<Payment> {
       issuer: offer_issuer,
       payerNote: invreq_payer_note
     }
+
+    const formatted = formatDecodedOffer(decoded)
+    decodedOffers$.next({ ...decodedOffers$.value, [offer_id]: formatted })
   }
 
   return {
